feat: export function computeHierarchicNodeLinks for sankey
